+++ packages/react-magma-docs/gatsby-node.js
@@ -23,4 +23,3 @@ exports.onCreateNode = ({ node, actions, getNode }) => {
     })
   }
 }
-